Got opaque movement working correctly; still need to add a few things: 1. put an event loop in mouse processing code to ignore EnterNotifys 2. Create a subwindow of frame to accomodate title and catch clicks 3. Ignore ButtonReleases unless the correct button is released 4. Perhaps work in window gravity, shouldn't be too hard 5. Clean up mouse interface to be able to bind buttons, etc.
Resizing will be much more complex than moving as we have to draw some
sort of indication of window size (not doing opaque resize, takes too
with most widget layout engines), and we have to honor size hints.

Might also do an XPeekEvent in the mouse code right before returning to
main event loop to eat up any stray EnterNotifys to accomodate the
move-without-focus thing.  Also need to experiment with this
move-without-focus to see if I really like it or not.